requirements.txt
absl-py==0.11.0
astor==0.8.1
beautifulsoup4==4.6.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
Django==2.2.6
gast==0.4.0
grpcio==1.35.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
JPype1==1.2.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
konlpy==0.5.2
lxml==4.6.2
Markdown==3.3.3
matplotlib==3.3.0
mock==4.0.3
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.1.1.26
Pillow==7.2.0
protobuf==3.14.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.20.3
scipy==1.6.0
six==1.15.0
sqlparse==0.4.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
zipp==3.4.0
